ultimo commit

+++ USAO_api/src/main/java/com/salesianostriana/dam/USAO_api/services/impl/ProductoServiceImpl.java
@@ -12,6 +12,7 @@ import com.salesianostriana.dam.USAO_api.services.StorageService;
 import com.salesianostriana.dam.USAO_api.users.dto.GetUserDto3;
 import com.salesianostriana.dam.USAO_api.users.dto.UserDtoConverter;
 import com.salesianostriana.dam.USAO_api.users.model.User;
+import com.salesianostriana.dam.USAO_api.users.model.UserRole;
 import com.salesianostriana.dam.USAO_api.users.repository.UserEntityRepository;
 import io.github.techgnious.exception.VideoException;
 import lombok.RequiredArgsConstructor;
@@ -101,25 +102,31 @@ public class ProductoServiceImpl {
                 String extension = StringUtils.getFilenameExtension(StringUtils.cleanPath(file.getOriginalFilename()));
                 List<String> allExtension = Arrays.asList("png", "gif", "jpg", "svg", "mp4");
 
+
                 if (!allExtension.contains(extension)) {
                     throw new UnsupportedMediaType(allExtension);
                 } else {
-                    String scale = StringUtils.cleanPath(String.valueOf(producto.get().getFileScale())).replace("http://localhost:8080/download/", "")
+                    /*String scale = StringUtils.cleanPath(String.valueOf(producto.get().getFileScale())).replace("http://localhost:8080/download/", "")
                             .replace("%20", " ");
-                    Path path = storageService.load(scale);
-                    String filename = StringUtils.cleanPath(String.valueOf(path)).replace("http://localhost:8080/download/", "")
-                            .replace("%20", " ");
-                    Path pathScalse = Paths.get(filename);
-                    storageService.deleteFile(pathScalse);
+
+                    if(scale.contains("download")){
+                        Path path = storageService.load(scale);
+                        String filename = StringUtils.cleanPath(String.valueOf(path)).replace("http://localhost:8080/download/", "")
+                                .replace("%20", " ");
+                        Path pathScalse = Paths.get(filename);
+                        storageService.deleteFile(pathScalse);
 
 
-                    String original = StringUtils.cleanPath(String.valueOf(producto.get().getFileOriginal())).replace("http://localhost:8080/download/", "")
-                            .replace("%20", " ");
-                    Path path2 = storageService.load(original);
-                    String filename2 = StringUtils.cleanPath(String.valueOf(path2)).replace("http://localhost:8080/download/", "")
-                            .replace("%20", " ");
-                    Path pathOriginal = Paths.get(filename2);
-                    storageService.deleteFile(pathOriginal);
+                        String original = StringUtils.cleanPath(String.valueOf(producto.get().getFileOriginal())).replace("http://localhost:8080/download/", "")
+                                .replace("%20", " ");
+                        Path path2 = storageService.load(original);
+                        String filename2 = StringUtils.cleanPath(String.valueOf(path2)).replace("http://localhost:8080/download/", "")
+                                .replace("%20", " ");
+                        Path pathOriginal = Paths.get(filename2);
+                        storageService.deleteFile(pathOriginal);
+                    }*/
+
+
 
 
                     String filenameOriginal = storageService.original(file);
@@ -160,11 +167,12 @@ public class ProductoServiceImpl {
             throw new SingleEntityNotFoundException(id.toString(), Producto.class);
         } else {
 
-            if(producto.get().getPropietario().getNick().equals(user.getNick())){
+            if(producto.get().getPropietario().getNick().equals(user.getNick()) || user.getRoles().equals(UserRole.ADMIN)){
                 String scale = StringUtils.cleanPath(String.valueOf(producto.get().getFileScale())).replace("https://usao-back.herokuapp.com/download/", "")
                         .replace("%20", " ");
 
                 if(!scale.contains("download")){
+
                     productoRepository.deleteById(id);
                 }else{
                     Path path = storageService.load(scale);
@@ -182,7 +190,6 @@ public class ProductoServiceImpl {
                     Path pathOriginal = Paths.get(filename2);
 
                     storageService.deleteFile(pathOriginal);
-
